guided project3
Plotly is a python graphing library which is used to make interactive, publication-quality graphs. It allows users to import, copy and paste, or stream data to be analyzed and visualized.In this project you will learn how to create beautiful visualizations using Plotly constructs. This guided project is for anyone who wants to learn data visualization or already in the data science field.